+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## HEAD
 
-- **Breaking change:** Remove support for Node 4.
+- **Breaking change:** Remove support for Node 4 and 6. Requires Node 8+.
 
 ## 5.2.1