+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD
 
+- Migrate from Flowtype to Typescript
 - **Breaking change:** Remove support for Node 4 and 6. Requires Node 8+.
 - **Breaking change:** Use named export `cosmiconfig`. (see example below)